Add `autocomplete=dummy` to most station detail fields
Since we're now using the same component for both contacted and logging station, and since we now have QTH profile, there's not really a good reason for browsers to try to autofill any of these fields. Tested with `autocomplete=off` which is more correct to HTML, but `dummy` seems to do a better job in Chrome.